# moltservice/gen/http/moltservice/server/encode_decode.go
// Code generated by goa v3.14.1, DO NOT EDIT.
//
// moltservice HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/cockroachdb/molt/moltservice/design -o ./moltservice

package server

import (
	"context"
	"io"
	"net/http"
	"strconv"

	moltservice "github.com/cockroachdb/molt/moltservice/gen/moltservice"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeCreateFetchTaskResponse returns an encoder for responses returned by
// the moltservice create_fetch_task endpoint.
func EncodeCreateFetchTaskResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(moltservice.FetchAttemptID)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeCreateFetchTaskRequest returns a decoder for requests sent to the
// moltservice create_fetch_task endpoint.
func DecodeCreateFetchTaskRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			body CreateFetchTaskRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateCreateFetchTaskRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewCreateFetchTaskCreateFetchPayload(&body)

		return payload, nil
	}
}

// EncodeGetFetchTasksResponse returns an encoder for responses returned by the
// moltservice get_fetch_tasks endpoint.
func EncodeGetFetchTasksResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*moltservice.FetchRun)
		enc := encoder(ctx, w)
		body := NewGetFetchTasksResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeGetSpecificFetchTaskResponse returns an encoder for responses returned
// by the moltservice get_specific_fetch_task endpoint.
func EncodeGetSpecificFetchTaskResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*moltservice.FetchRunDetailed)
		enc := encoder(ctx, w)
		body := NewGetSpecificFetchTaskResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetSpecificFetchTaskRequest returns a decoder for requests sent to the
// moltservice get_specific_fetch_task endpoint.
func DecodeGetSpecificFetchTaskRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			id  int
			err error

			params = mux.Vars(r)
		)
		{
			idRaw := params["id"]
			v, err2 := strconv.ParseInt(idRaw, 10, strconv.IntSize)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("id", idRaw, "integer"))
			}
			id = int(v)
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetSpecificFetchTaskPayload(id)

		return payload, nil
	}
}

// EncodeCreateVerifyTaskFromFetchResponse returns an encoder for responses
// returned by the moltservice create_verify_task_from_fetch endpoint.
func EncodeCreateVerifyTaskFromFetchResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(moltservice.VerifyAttemptID)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeCreateVerifyTaskFromFetchRequest returns a decoder for requests sent
// to the moltservice create_verify_task_from_fetch endpoint.
func DecodeCreateVerifyTaskFromFetchRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			id  int
			err error

			params = mux.Vars(r)
		)
		{
			idRaw := params["id"]
			v, err2 := strconv.ParseInt(idRaw, 10, strconv.IntSize)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("id", idRaw, "integer"))
			}
			id = int(v)
		}
		if err != nil {
			return nil, err
		}
		payload := NewCreateVerifyTaskFromFetchPayload(id)

		return payload, nil
	}
}

// marshalMoltserviceFetchRunToFetchRunResponse builds a value of type
// *FetchRunResponse from a value of type *moltservice.FetchRun.
func marshalMoltserviceFetchRunToFetchRunResponse(v *moltservice.FetchRun) *FetchRunResponse {
	res := &FetchRunResponse{
		ID:         v.ID,
		Name:       v.Name,
		Status:     v.Status,
		StartedAt:  v.StartedAt,
		FinishedAt: v.FinishedAt,
	}

	return res
}

// marshalMoltserviceFetchStatsDetailedToFetchStatsDetailedResponseBody builds
// a value of type *FetchStatsDetailedResponseBody from a value of type
// *moltservice.FetchStatsDetailed.
func marshalMoltserviceFetchStatsDetailedToFetchStatsDetailedResponseBody(v *moltservice.FetchStatsDetailed) *FetchStatsDetailedResponseBody {
	if v == nil {
		return nil
	}
	res := &FetchStatsDetailedResponseBody{
		PercentComplete:  v.PercentComplete,
		NumErrors:        v.NumErrors,
		NumTables:        v.NumTables,
		NumRows:          v.NumRows,
		NetDurationMs:    v.NetDurationMs,
		ImportDurationMs: v.ImportDurationMs,
		ExportDurationMs: v.ExportDurationMs,
		CdcCursor:        v.CdcCursor,
	}

	return res
}

// marshalMoltserviceLogToLogResponseBody builds a value of type
// *LogResponseBody from a value of type *moltservice.Log.
func marshalMoltserviceLogToLogResponseBody(v *moltservice.Log) *LogResponseBody {
	res := &LogResponseBody{
		Timestamp: v.Timestamp,
		Level:     v.Level,
		Message:   v.Message,
	}

	return res
}