# moltservice/gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"MOLT Service","description":"Service for coordinating between clients (CLI + Web UI) and MOLT tooling","version":"1.0"},"servers":[{"url":"http://localhost:4500"}],"paths":{"/api/v1/fetch":{"get":{"tags":["moltservice"],"summary":"get_fetch_tasks moltservice","operationId":"moltservice#get_fetch_tasks","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FetchRun"},"example":[{"finished_at":1704233521,"id":1704233521,"name":"jyang pg to crdb","started_at":1704233519,"status":"IN_PROGRESS"},{"finished_at":1704233521,"id":1704233521,"name":"jyang pg to crdb","started_at":1704233519,"status":"IN_PROGRESS"},{"finished_at":1704233521,"id":1704233521,"name":"jyang pg to crdb","started_at":1704233519,"status":"IN_PROGRESS"},{"finished_at":1704233521,"id":1704233521,"name":"jyang pg to crdb","started_at":1704233519,"status":"IN_PROGRESS"}]},"example":[{"finished_at":1704233521,"id":1704233521,"name":"jyang pg to crdb","started_at":1704233519,"status":"IN_PROGRESS"},{"finished_at":1704233521,"id":1704233521,"name":"jyang pg to crdb","started_at":1704233519,"status":"IN_PROGRESS"}]}}}}},"post":{"tags":["moltservice"],"summary":"create_fetch_task moltservice","operationId":"moltservice#create_fetch_task","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFetchTaskRequestBody"},"example":{"bucket_name":"http://localhost:5000","bucket_path":"fetch/export","cleanup_intermediary_store":true,"compression":"none","local_path":"/usr/Documents/fetch","local_path_crdb_address":"http://localhost:5000","local_path_listen_address":"http://localhost:5000","log_file":"task.log","mode":"DIRECT_COPY","name":"rluu pg to cockroach","num_batch_rows_export":100000,"num_concurrent_tables":4,"num_flush_bytes":2000000,"num_flush_rows":200000,"pg_drop_slot":false,"pg_logical_plugin":"","pg_logical_slot_name":"","source_conn":"postgres://postgres:postgres@localhost:5432/molt?sslmode=disable","store":"Local","target_conn":"postgres://root@localhost:26257/defaultdb?sslmode=disable","truncate":true}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":3120370108600835844,"format":"int64"},"example":1788448450945662325}}}}}},"/api/v1/fetch/{id}":{"get":{"tags":["moltservice"],"summary":"get_specific_fetch_task moltservice","operationId":"moltservice#get_specific_fetch_task","parameters":[{"name":"id","in":"path","description":"id for the fetch task","required":true,"schema":{"type":"integer","description":"id for the fetch task","example":4061375660015501511,"format":"int64"},"example":5809114746007290805}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FetchRunDetailed"},"example":{"finished_at":1704233521,"id":1704233521,"logs":[{"level":"INFO","message":"This is a log message","timestamp":1704233519},{"level":"INFO","message":"This is a log message","timestamp":1704233519},{"level":"INFO","message":"This is a log message","timestamp":1704233519}],"name":"jyang pg to crdb","started_at":1704233519,"stats":{"cdc_cursor":"0/3F3E0B8","export_duration_ms":100000,"import_duration_ms":100000,"net_duration_ms":100000,"num_errors":0,"num_rows":100000,"num_tables":5,"percent_complete":"jyang pg to crdb"},"status":"IN_PROGRESS"}}}}}}},"/api/v1/fetch/{id}/verify":{"post":{"tags":["moltservice"],"summary":"create_verify_task_from_fetch moltservice","operationId":"moltservice#create_verify_task_from_fetch","parameters":[{"name":"id","in":"path","description":"id for the fetch task","required":true,"schema":{"type":"integer","description":"id for the fetch task","example":8529004787339204778,"format":"int64"},"example":434544346317199182}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":2667229789776596173,"format":"int64"},"example":7153356925827642199}}}}}},"/docs.html":{"get":{"tags":["moltservice"],"summary":"Download ./assets/docs.html","operationId":"moltservice#/docs.html","responses":{"200":{"description":"File downloaded"}}}},"/openapi.json":{"get":{"tags":["moltservice"],"summary":"Download ./gen/http/openapi.json","operationId":"moltservice#/openapi.json","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"CreateFetchTaskRequestBody":{"type":"object","properties":{"bucket_name":{"type":"string","description":"the local address CRDB will use to access the import files","example":"http://localhost:5000"},"bucket_path":{"type":"string","description":"the sub-path within the bucket to write the export files","example":"fetch/export"},"cleanup_intermediary_store":{"type":"boolean","description":"whether the intermediate store should be cleaned up after the fetch task","example":true},"compression":{"type":"string","description":"compression type","example":"none","enum":["gzip","none","default"]},"local_path":{"type":"string","description":"the absolute or relative path to write export files","example":"/usr/Documents/fetch"},"local_path_crdb_address":{"type":"string","description":"the local address CRDB will use to access the import files","example":"http://localhost:5000"},"local_path_listen_address":{"type":"string","description":"the local address where the file server will be spun up","example":"http://localhost:5000"},"log_file":{"type":"string","description":"if specified, writes task execution logs to a file and stdout; otherwise, just writes to stdout","example":"task.log"},"mode":{"type":"string","description":"Mode of operation for fetch","example":"DIRECT_COPY","enum":["IMPORT_INTO","COPY_FROM","DIRECT_COPY"]},"name":{"type":"string","description":"the name of the fetch run","example":"rluu pg to cockroach"},"num_batch_rows_export":{"type":"integer","description":"number of rows to export at a given time for each iteration; tune this so that you get most out of CPU and can batch the most data together","example":100000,"format":"int64"},"num_concurrent_tables":{"type":"integer","description":"number of tables to process at the same time; this is usually sized based on number of CPUs","example":4,"format":"int64"},"num_flush_bytes":{"type":"integer","description":"number of bytes for the export before data is flushed to the disk","example":2000000,"format":"int64"},"num_flush_rows":{"type":"integer","description":"number of rows for the export before data is flushed to the disk (persisted)","example":200000,"format":"int64"},"pg_drop_slot":{"type":"boolean","description":"if set and exists, drops the existing replication slot","example":false},"pg_logical_plugin":{"type":"string","description":"name for pg replication plugin","example":""},"pg_logical_slot_name":{"type":"string","description":"name for pg replication slot","example":""},"source_conn":{"type":"string","description":"source database connection string","example":"postgres://postgres:postgres@localhost:5432/molt?sslmode=disable"},"store":{"type":"string","description":"Type of intermediary store","example":"Local","enum":["None","AWS","GCP","Local"]},"target_conn":{"type":"string","description":"target database connection string","example":"postgres://root@localhost:26257/defaultdb?sslmode=disable"},"truncate":{"type":"boolean","description":"if specified, truncates the target tables before running the data load","example":true}},"example":{"bucket_name":"http://localhost:5000","bucket_path":"fetch/export","cleanup_intermediary_store":true,"compression":"none","local_path":"/usr/Documents/fetch","local_path_crdb_address":"http://localhost:5000","local_path_listen_address":"http://localhost:5000","log_file":"task.log","mode":"DIRECT_COPY","name":"rluu pg to cockroach","num_batch_rows_export":100000,"num_concurrent_tables":4,"num_flush_bytes":2000000,"num_flush_rows":200000,"pg_drop_slot":false,"pg_logical_plugin":"","pg_logical_slot_name":"","source_conn":"postgres://postgres:postgres@localhost:5432/molt?sslmode=disable","store":"Local","target_conn":"postgres://root@localhost:26257/defaultdb?sslmode=disable","truncate":true},"required":["source_conn","target_conn","mode","store","cleanup_intermediary_store","local_path","local_path_listen_address","local_path_crdb_address","bucket_name","bucket_path","log_file","truncate","compression","num_flush_rows","num_flush_bytes","num_concurrent_tables","num_batch_rows_export","pg_logical_slot_name","pg_logical_plugin","pg_drop_slot","name"]},"FetchRun":{"type":"object","properties":{"finished_at":{"type":"integer","description":"finished at time","example":1704233521,"format":"int64"},"id":{"type":"integer","description":"ID of the run","example":1704233521,"format":"int64"},"name":{"type":"string","description":"name of the fetch run","example":"jyang pg to crdb"},"started_at":{"type":"integer","description":"started at time","example":1704233519,"format":"int64"},"status":{"type":"string","description":"status of the fetch run","example":"IN_PROGRESS","enum":["IN_PROGRESS","SUCCESS","FAILURE"]}},"example":{"finished_at":1704233521,"id":1704233521,"name":"jyang pg to crdb","started_at":1704233519,"status":"IN_PROGRESS"},"required":["id","name","status","started_at","finished_at"]},"FetchRunDetailed":{"type":"object","properties":{"finished_at":{"type":"integer","description":"finished at time","example":1704233521,"format":"int64"},"id":{"type":"integer","description":"ID of the run","example":1704233521,"format":"int64"},"logs":{"type":"array","items":{"$ref":"#/components/schemas/Log"},"description":"logs for fetch run","example":[{"level":"INFO","message":"This is a log message","timestamp":1704233519},{"level":"INFO","message":"This is a log message","timestamp":1704233519}]},"name":{"type":"string","description":"name of the fetch run","example":"jyang pg to crdb"},"started_at":{"type":"integer","description":"started at time","example":1704233519,"format":"int64"},"stats":{"$ref":"#/components/schemas/FetchStatsDetailed"},"status":{"type":"string","description":"status of the fetch run","example":"IN_PROGRESS","enum":["IN_PROGRESS","SUCCESS","FAILURE"]}},"example":{"finished_at":1704233521,"id":1704233521,"logs":[{"level":"INFO","message":"This is a log message","timestamp":1704233519},{"level":"INFO","message":"This is a log message","timestamp":1704233519}],"name":"jyang pg to crdb","started_at":1704233519,"stats":{"cdc_cursor":"0/3F3E0B8","export_duration_ms":100000,"import_duration_ms":100000,"net_duration_ms":100000,"num_errors":0,"num_rows":100000,"num_tables":5,"percent_complete":"jyang pg to crdb"},"status":"IN_PROGRESS"},"required":["id","name","status","started_at","finished_at","logs"]},"FetchStatsDetailed":{"type":"object","properties":{"cdc_cursor":{"type":"string","description":"CDC cursor","example":"0/3F3E0B8"},"export_duration_ms":{"type":"number","description":"export duration in milliseconds","example":100000,"format":"double"},"import_duration_ms":{"type":"number","description":"import duration in milliseconds","example":100000,"format":"double"},"net_duration_ms":{"type":"number","description":"net duration in milliseconds","example":100000,"format":"double"},"num_errors":{"type":"integer","description":"number of errors processed","example":0,"format":"int64"},"num_rows":{"type":"integer","description":"number of rows","example":100000,"format":"int64"},"num_tables":{"type":"integer","description":"number of tables processed","example":5,"format":"int64"},"percent_complete":{"type":"string","description":"percentage complete of fetch run","example":"jyang pg to crdb"}},"example":{"cdc_cursor":"0/3F3E0B8","export_duration_ms":100000,"import_duration_ms":100000,"net_duration_ms":100000,"num_errors":0,"num_rows":100000,"num_tables":5,"percent_complete":"jyang pg to crdb"},"required":["percent_complete","num_errors","num_tables","num_rows","net_duration_ms","import_duration_ms","export_duration_ms","cdc_cursor"]},"Log":{"type":"object","properties":{"level":{"type":"string","description":"level for logging","example":"INFO"},"message":{"type":"string","description":"message for the logging","example":"This is a log message"},"timestamp":{"type":"integer","description":"timestamp of log","example":1704233519,"format":"int64"}},"example":{"level":"INFO","message":"This is a log message","timestamp":1704233519},"required":["timestamp","level","message"]}}},"tags":[{"name":"moltservice","description":"MOLT service performs operations using MOLT tooling"}]}